fix: transform xml to json
create new file to increase readability of code
