+++ lwm2m/typeboxDefinition.ts
@@ -1,7 +1,6 @@
 import { Type, type TSchema } from '@sinclair/typebox'
-import xml2js from 'xml2js'
-import { readFile } from 'node:fs/promises'
-import path from 'node:path'
+import { fromXML2JSON } from './fromXML2JSON.js';
+
 
 /**
  * Create Typebox definition for object 14201
@@ -75,30 +74,7 @@ export const typeboxDefinition = async (id: number): Promise<TSchema> => {
 	)
 }
 
-/**
- * JSON representation of an XML object
- */
-type jsonObject = {
-	Name: string[]
-	Description1: string[]
-	ObjectID: string[]
-	ObjectURN: string[]
-	LWM2MVersion: string[]
-	ObjectVersion: string[]
-	MultipleInstances: ['Single'] | ['Multiple']
-	Mandatory: ['Optional'] | ['Mandatory']
-	Resources: [Record<string, unknown>]
-	Description2: string[]
+const writeTypeboxDefinition = () => {
+
 }
 
-/**
- * From XML to JSON
- */
-const fromXML2JSON = async (id: number): Promise<jsonObject> => {
-	const baseDir = process.cwd()
-	const subDir = (...tree: string[]): string => path.join(baseDir, ...tree)
-	const jsonObject = await xml2js.parseStringPromise(
-		await readFile(subDir('lwm2m', `${id}.xml`), 'utf-8'),
-	)
-	return jsonObject.LWM2M.Object[0]
-}
